Important Disclosure  - No Registration
SaveCoin Tokens, when offered, will not be registered under the US Securities Act of 1933 (the "securities Act"), the securities laws of any state, or any other jurisdiction, nor is such registration contemplated. SaveCoin Tokens will be offered and sold pursuant to one or more exemptions from the registration requirements of the Securities Act and any other applicable jurisdiction.

In all jurisdictions, the offer to sell and solicitation to buy SaveCoin Tokens is directed solely to qualified accredited investors ("Qualified Person").  This website does not constitute an offer, distribution, solicitation, or marketing to any non-Qualified Person (and is not an offering to the retail public in any jurisdiction where such offering is unlawful).  You should disregard this if you are a non-qualified person.